This revision adds the supporting research on Design Patterns using Java. The documents were used during the research to construct a few Java implementations of the theory.
This revision includes different implementation of Go4 Design Patterns in Java.
I started this implementation as a way to practive during the first days after I
started in the Graduate Program in Software Test Engineering at Motorola and
Center of Informatics from the Federal University of Pernambuco (UFPE), Brazil.
I worked as a Software Test Engineer and did research in Software Engineering in
Test. Finally, this revision adds the tests for the behavioral Go4 Design Patterns developed
during the graduate program in Software Engineering in Test at the Federal
University of Pernambuco.