This revision adds the working version of the computation of the Feature normalization of data sets using Octave for the Stanford Machine Learning class. This is an extra-credit exercise submitted successfully. http://www.ml-class.org.
As described in the documentation, the feature normalization can be
calculated using the mean and standard deviation functions from 
octave on the training example. The manipulation of the matrix elements
can be used in an array of index for a given feature, and the
computations can be performed using the element-wise arithmetics.